Dear god, vi is annoying.
Ahem.

So this is a patch to combine the FAQ and the About pages into about.php. Why?
Because when i ran through the site a couple of times, I immediately followed
the "how this site works" link to /about, but had lots of questions that weren't
answered by /about. My first instinct was to add a link to /faq, but then I
realised that was just a bit silly. So I kept the good bits of about.php, added
in the questions from faq.php in a sensible order, and hey presto.

There's a related patch coming through in a moment, which modifies the footer to
point at /about instead of /faq.

What think ye? Is this a good way to combine them? Do they need to be combined?